Refactor internal event API calls to use DTO + rename internal client
